requirements.txt
# Core dependencies
tensorflow>=2.15.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
pillow>=10.0.0

# GUI
tkinterweb>=3.24.0; sys_platform == "win32"

# Optional utilities
pandas>=2.0.0

# Environment
python-dotenv>=1.0.0
